working on delta indexing
