+++ meme_search/style.css
@@ -3,12 +3,12 @@ body {
     background-color: #4F8BF9;
 }
 
-.stButton>button {
+/* .stButton>button {
     color: #4F8BF9;
     border-radius: 50%;
     height: 3em;
     width: 3em;
-}
+} */
 
 .stTextInput>div>div>input {
     color: #4F8BF9;
